Fix type error in thin-threshold dedup query
CASE branches with untyped params ($4/$6) defaulted to text, so the
outer `>` comparison against a double-precision similarity value failed
with "operator does not exist: double precision > text" — killed the
whole fetch run on the first article that reached findAndAssignGroup.
Cast both branches to float8 explicitly.

+++ src/lib/grouping.ts
@@ -54,7 +54,7 @@ export async function findAndAssignGroup(
        AND ($2::int IS NULL OR source_id != $2)
        AND created_at > now() - interval '${options.windowHours} hours'
        AND 1 - (embedding <=> $3::vector) > (
-         CASE WHEN $5::bool OR full_description IS NULL THEN $6 ELSE $4 END
+         CASE WHEN $5::bool OR full_description IS NULL THEN $6::float8 ELSE $4::float8 END
        )
      ORDER BY embedding <=> $3::vector
      LIMIT 1`,
